requirements.txt
yfinance==0.1.70
alpaca-trade-api==1.5.0
datetime==4.4
python-dotenv==0.19.2
numpy==1.21.5
pandas==1.3.5
tweepy==4.10.0
python-coinmarketcap==0.3